All routes protected

+++ routes/bookmarks.js
@@ -1,19 +1,18 @@
 const express = require('express');
 const {Bookmark} = require('../database');
 var router = express.Router();
-
 const passport = require('passport');
 const passportJWT = require("passport-jwt");
 const ExtractJWT = passportJWT.ExtractJwt;
+const middleware = require('../middleware/middlewareUser.js');
 
-router.post('/new', (req, res) => {
-    console.log(new Buffer(req.headers.authorization.split(' ')[1].split('.')[1], 'base64').toString());
+router.post('/new', middleware.extractUserIdBody, (req, res) => {
     Bookmark.create(req.body)
         .then(bookmark => res.json(bookmark));
 });
 
-router.delete('/delete', (req,res) => {
-    Bookmark.destroy({ where: {id : req.query.id}} )
+router.delete('/delete', middleware.extractUserIdQuery, (req,res) => {
+    Bookmark.destroy({ where: {id : req.query.id, userId : req.query.userId}} )
         .then(res.json(true));
 });
